Add translation system and a language changer to the front page
As the reader may guess from the commit time, this is the result of
what we in the business call "jumping straight into the rabbit hole."

When I started writing a template index.html for some basic forum
functionality, I realized that a translation system would be nice to
have, before writing the entire system in English and having no easy
way to translate it afterwards. This led me to learn about GNU
gettext, what /usr/share/locale/ contains, what .po and .mo files are,
and how that all can be used in Jinja. Truly, an excellent evening of
learning.